When you add a node, the initial number is set to the next available. This occurs whether or not you specify a node number on the command line.
If you don't specify a number, o2cb_ctl continues and all is fine.

If you do specify a number, o2cb_ctl goes to overwrite the default
number with your new number.  However, it also checks to see if there is
an existing node with that number:

    tmpnode = o2cb_get_node_by_num(cluster, num)
    if (tmpnode)
        error("Already exists!")

BUT, we've just added are new skeleton node, and its default number is
the same as the number you are asking for -- the next in the sequence.
So we are seeing the node we are adding as already existing!

The fix is to check for an existing node that isn't our new node.

    tmpnode = o2cb_get_node_by_num(cluster, num)
    if (tmpnode && (tmpnode != newnode))
        error("Already exists!")

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>

Original-revision: 1231