simplify o2fsck_bitcount and fix it to cover all data.
Tao Ma noticed that previously o2fsck_bitcount() forgot to increment its
'bytes' pointer and so only counted the first 32bit word of data given.

We replace the incomprehensible computational gymnastics with a trivial loop
that uses a 4 bit lookup table that contains the precomputed number of bits set
in each index value.  A human can audit this as being obviously correct in a
number of minutes.

In trivial testing it returns the same value as the previous implementation
after the latter was fixed to walk the entire buffer.  As a bonus, it uses
about half as many cycles as the previous implementation when operating on
cached data on modern processors.

Signed-off-by: Zach Brown <zach.brown@oracle.com>
Signed-off-by: Zach Brown <zach.brown@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>

Original-revision: 1233