ocfs2-tools: Set right start position for bitmap finding.
In write_cluster_alloc, every bits in global bitmap will
be checked according to the real allocation. When some
clusters are found to be cleared from the bitmap, they will
not be cleared except the first one due to the wrong start
searching position. So change it to the right one. 

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Zach Brown <zach.brown@oracle.com>

Original-revision: 1236