ocfs2-tools: Add file type when create dot or dotdot.
when fsck.ocfs2 finds some error and  wants to add a new dot or dotdot
to the directory, it should set the file type accordingly.
Or since inode is equal to zero, the file type will not be set during this
run of fsck.ocfs2. Another fsck.ocfs2 is needed to set the right file type.
It is obviously not good for a user.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>

Original-revision: 1265