Add the mechanism of bitmap scan in fsck.ocfs2.
When the bg_free_bits_count of a group descriptor is corrupted, fsck can't
fix it now because of the simple check. ct met with this problem. See
http://oss.oracle.com/bugzilla/show_bug.cgi?id=841.

So now fsck.ocfs2 wil scan the bitmap, find the count of real free bits,
compare and fix the bg_free_bits_count stored in the group descriptor.
This bug should be fixed.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>

Original-revision: 1290