Start and stop the o2cb_controld and ocfs2_controld daemons. This includes
making sure our cluster name matches the cman one, as well as not stopping
when a heartbeat region is live.

We still need to clean up the signal handling and exit code for the control
daemons.  It works right now, but it's ugly.

Also, teach load_status to show the heartbeat driver.

Original-revision: 1385