ocfs2_controld now handles signals and exit cleanly.
Add a proper signal handler that feeds the signals to a pipe.  This way
the main loop can read the signal via POLLIN.  The signal processing
function knows to exit if there are no mounts and stay alive if there are.

When the main loop is broken due to a signal or a cman disconnect, we
now clean up in our exit path.  Any existing mounts are removed from
user heartbeat.  This can cause a self-fence if there are live mounts, but
that only happens on a very abnormal exit (SEGV, cman exiting) where we
can't stay alive safely anyway.

The functionality for cleaning up a user heartbeat region is moved out
of do_terminate() so it can be used by the exit code.

Original-revision: 1398