Export o2cb_create/remove_heartbeat_region().
The low-level o2cb_create/remove_heartbeat_region() functions are needed by
ocfs2_controld to create user regions.  Heartbeat2 might want it as well.
Everyone else should be using the join/leave API, which will communicate
with the proper actors to get the configfs stuff munged.

The functions correctly handle creating disk regions for the o2cb stack and
plain directory regions for the user heartbeat stack.

o2cb.h is also reorganized a little to comment on functions that are for
low-level use and those that are more useful for high-level programs.

Original-revision: 1400