Here is an I/O cache backing unix_io.c in libocfs2.  It's a simple read cache.  Writes are write-through and synchronous.
The user specifies the cache size in bytes or blocks.  When they call
io_read_block() or io_write_block(), the block is looked up in the
cache.  If it exists, a read merely copies the data to the user buffer,
while a write copies the user data to the cache buffer, then syncs the
buffer to disk (write-through).

If the block doesn't exist in the cache, the least recently used cache
block is stolen and filled as above.

The cache is not automatically created during ocfs2_open().  A process
that wants caching needs to enable it after taking the appropriate
locking steps.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>

Original-revision: 1407