configure: Detect systems where we cannot statically build o2cb_ctl.
Fedora 7 no longer provides static libraries for GLib 2.  Thus, we
cannot statically link o2cb_ctl on that platform.  While the
--enable-dynamic-ctl option works, this is really something we should
detect.

If --enable/disable-dynamic-ctl is not specified, configure will try to
link a static program with GLib 2.  If this works, o2cb_ctl will be
built static, otherwise it will be built dynamic.  We honor
--enable/disable-dynamic-ctl when it is specified.  Caveat emptor.

While we're here, we now error if GLib 2 is not found.  We require it
for o2cb_ctl.  As GLib 2 it now required, we no longer conditionally
build ocfs2cdsl.  That test is cleaned up too.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>