control daemons: Make sure they understand each other's presence.
Basically, if o2cb_controld dies, ocfs2_controld has to die, as node
info can't get updated.  Conversely, if ocfs2_controld is alive,
o2cb_controld should try very hard not to die.

To do this, o2cb_controld listens for connections on a Unix socket.
ocfs2_controld connects to the socket.  While there are connections on
the socket, o2cb_controld will not exit on a signal.

If killed with -9, or if cman goes away, o2cb_controld can still go
away.  When this happens, ocfs2_controld will notice and exit itself.
This will take out the node if there are ocfs2 mounts, but that is the
right thing to do.

Signed-off-by: Joel Becker <joel.becker@oracle.com>