tunefs.ocfs2: Don't set incompat flags unless we're really changing slots
opts.num_slots is initialized to zero, which is always less than max_slots
on a valid ocfs2 file system. If tunefs.ocfs2 is called at all for anything,
even without "-N" the test for setting the removing-slots incompat flag
succeeds because it's not explicitely checking for zero.

This means that simple operations like changing mount type leave the
incompat bit set in the super block, rendering the file system unmountable
until fsck.ocfs2 is run.

Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>