Modify the check in io_cache_disconnect.
In the io_cache rb-tree, we may want to erase a cached block
which is the root of the whole tree. In this scenario, the rb_parent
will be NULL. So this io_cache will not be removed by the original
io_cache_disconnect and lead to the "assert" problem when we try
io_cache_insert by using the same icb.

Instead, the icb->icb_blkno will be checked when we want to remove
the specified io_cache. The icb->icb_blkno is set to to UINT64_MAX
when it is removed and we also initialize it to UINT64_MAX in
io_init_cache.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>