Add "clear sparse" feature for ocfs2-tools.
With a new formatted volume, we can use "--fs-features=nosparse"
in tunefs.ocfs2 to clear the sparse flag and let it to be used by
old ocfs2 kernel. By this option, all the holes will be allocated
with spaces and all the unwritten extents will be removed with
their contents emptied with zero.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>