Fix a wrong usage of "startblk" during offline resize and backup super.
During offline resize, opts.num_blocks indicates the block count for
the ocfs2 volume, not the increased block count. But update_backup_super
use the wrong meaning by mistake. So when we increase the block into a
a number which isn't aligned with cluster size, startblk will overflow
since fs->fs_blocks will be lower than num_blocks and no backup super
blocks will be allocated since ocfs2_set_backup_super will never be
called.

So record the old_blocks when we do offline resize and that is the
really start offset for the new blocks.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>