tunefs: add ability to turn on unwritten extents
If sparse file support is enabled, turning on unwritten extents is just a
matter of writing the ro_compat flag. If the user asks for unwritten and the
file system doesn't support sparse files, we turn on sparse 1st.

The feature set code got a little smarter as a result of this patch - we can
now just use a set of bitfields to represent which features tunefs knows how
to set or clear.

Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>