ocfs2-tools: Import the latest ocfs2_fs.h
The extended slot map is defined in the latest version of ocfs2_fs.h.  In
addition, it defines a proper data structure for the old slot map.

This is a straight import from te kernel sources.  The only modification
is to remove OCFS2_FEATURE_INCOMPAT_EXTENDED_SLOT_MAP from the SUPPORTED
bitmask.  It will be added back when the tools support is complete.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>