ocfs2-tools: Remove open-coded slot map access.
Anywhere the tools need to access the slot map, they open-code access.
Introduce a set of functions in libocfs2 to manage the slot map.  The
best is the load/store functions, which provide an in-memory structure
to isolate users from the specifics of the format.

We clean up the open-coded places as well.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>