mkfs.ocfs2: Support the extended slot map.
mkfs.ocfs2 had an open-coded function for writing an initial slot map.
It is removed in favor of ocfs2_format_slot_map(), which can handle
both the old and new formats.  The feature string "extended-slotmap" is
added to trigger the new format.

The format of the slot map is moved to where we can use ocfs2_open() before
calling ocfs2_format_slot_map().  It's perfectly safe to do there, as no
one can mount without a slot map.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>