ocfs2_controld: Leave groups
The daemon now handles leaving a group.  When a umount request is
recieved, it will leave the matching group.  When we are notified that
we've left, we'll tell the mount code.

In turn, the mount code handles normal and abnormal leaves.  In the
abnormal case, it will drop all nodes and then drop the group.

This allows us to reinstate bail_on_mounts() for the unplanned daemon
exit.  It just relies on simulating abnormal leaves.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>