o2cb.init: Eliminate the single-path load_actions scheme.
The o2cb.init script had an array of actions that are run at load time.
Upon unload, they are run in reverse.  This worked great for a single
backing cluster stack, but does not work well when we want to support
multiple schemes.  The steps are now broken out.

The mechanism for loading modules and pseudo-filesystems could not handle
compiled-in drivers.  Because depmod works great with loading modules for
filesystems, the code can just mount a filesystem and let depmod do the
rest.

o2cb.init is moved to o2cb.init.sh.  This allows us to propagate
variables to the init script when the time comes.  This involves adding
ALL_RULES to Postamble.make, something that already has happened in
Makebo.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>