ocfs2_controld: Add mountgroups and mount client communication.
The daemon now accepts the MOUNT, MRESULT, and UNMOUNT messages as
specified in the protocol.  It builds mountgroup structures as well as
the associated mountpoints.  It properly handles secondary mountpoints
as well as all but one client death scenario.

Currently it fakes joining and leaving groups.  That's the next step.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>