fsck.ocfs2: Fail to do normal block check when the file has tree depth > 1.
For a file which has l_tree_depth > 1, fsck.ocfs2 fail to check it. And even worse,
it writes the wrong value to i_last_eb_blk and return no error. So next time when
the volume is mounted and this file is written, fs will be made read-only. This is
caused by 2 bugs.

1. ocfs2_extent_iterate_inode intializes the walk-through extent block buffer pointer
   to a wrong value.
2. We write back i_last_eb_blk regardless of the return value.

So fix them.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Mark Fasheh <mark.fasheh@oracle.com>