Add online resize support in tunefs.ocfs2.
User can do offline resize using tunefs.ocfs2 when a volume isn't
mounted. Now the support for online resize is added.

Please note that the node where online resize goes must already
has the volume mounted. We don't mount it behind the user and the
operation would fail if we find it isn't mounted. As for other
nodes, we don't care whether the volume is mounted or not.

global_bitmap, super block and all the backups will be updated
in the kernel. And if super block or backup's update fails, we
just output some error message in dmesg and continue the work.

The whole process is derived from ext3 and divided into 2 steps:
1. OCFS2_IOC_GROUP_EXTEND: We will extend the last group first
   if it isn't full. All the work is done in kernel.
2. For every new group,  call OCFS2_IOC_GROUP_ADD to add them
   one by one. We initialize the new group descriptor in userspace,
   and let the kernel update the global_bitap etc.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>