o2cb: Introduce the cluster info struct.
The filesystem now has a field on disk called ocfs2_cluster_info.  It
contains the name of the cluster stack and the name of the cluster.  It
is only valid if the USERSPACE_STACK incompat bit is set.  If the bit is
not set, the filesystem expects the classic stack.

libocfs2 gains a routine to read this field.  It fills in a struct
o2cb_cluster_desc.  If the classic stack is in use, the structure is
filled in with NULLs.

Likewise, o2cb_*_group_*() take the o2cb_cluster_desc structure rather
than a cluster name.  This allows them to check the stack expected by
the filesystem with the stack actually in use.  If a userspace stack,
the name will be passed over the wire to the control daemon.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>