fsck.ocfs2: Fix cluster information in fsck.ocfs2
If the running cluster differs from the cluster information written in
the superblock, fsck.ocfs2 will not be able to lock the filesystem.  fsck
can now replace the information in the superblock with the running
configuration, allowing it to proceed with cluster locking.

The actual code to install the cluster configuration is placed in libocfs2
so tha tunefs can use it too.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>