Adds o2image header and supporting functions
o2image - An ocfs2 tool to save critical ocfs2 filesystem metadata to a
specified image-file. Image-file may be examined using debugfs.ocfs2 or
may be used to restore using o2image. Image-file can be of two formats:

1. Packed - This format(default) contains a o2image header, packed
metadata blocks and a bitmap.
2. raw    - A raw image is a sparse file containing the metadata blocks.

Usage: o2image [-rI] <device> <imagefile>

Packed format contains bitmap towards the end of the image-file. Each bit in
the bitmap represents a block in the filesystem.

When the packed image is opened using o2image or debugfs.ocfs2, bitmap is
loaded into memory and used to map disk blocks to image blocks.

Raw image is a sparse file containing metadata blocks at the same offset as
the filesystem.

debugfs.ocfs2 is modified to detect image-file when the image-file is
specified with -i option.

This patch adds o2image header files and supporting functions to read
write to image-file

Signed-off-by: Srinivas Eeda <srinivas.eeda@oracle.com>