ocfs2_hb_ctl: Use ocfs2_hb_ctl again for o2cb umount(2).
We've determined that mountpoints cannot be safely tracked from
userspace.  The only safe place to leave a filesystem group is in
put_super(7).  We used to do this via ocfs2_hb_ctl, and we're going to
do so again.

This change does three things:

1) Puts ocfs2_hb_ctl(8) back in /proc/sys/fs/ocfs2/nm/hb_ctl_path
2) Teaches ocfs2_hb_ctl(8) to do nothing when the cluster stack isn't o2cb
3) Teaches umount.ocfs2(8) to skip cluster interaction when the cluster
   stack *is* o2cb.

Thus, ocfs2_hb_ctl(8) is used for o2cb clusters and umount.ocfs2(8) is
used for all other cluster stacks.  Future changes will add support for
all cluster stacks to ocfs2_hb_ctk(8), removing the need for
umount.ocfs2(8).

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>