libocfs2: ocfs2_write_super() should update all superblocks
Normal operations should keep the backup superblocks up to date.  Thus,
ocfs2_write_super() is modified to refresh the backups.
ocfs2_write_primary_super() is introduced for operations that want to
leave the backups untouched.

The backup superblock APIs are slightly modified for clarity.  In
particular, ocfs2_refresh_backup_supers() is introduced to refresh all
backups.  It does nothing if backup superblocks are not enabled.  The
old API are kept but deprecated.

The rest of the tools now use the new APIs.  This includes using
ocfs2_write_primary_super() where necessary.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>