tunefs.ocfs2: Fix bit count error in offline resize.
commit cee1918350761300cb298248154e3b514e7a44aa (Add online resize support in
tunefs.ocfs2) introduced a bug in offline resize.  The global bitmap's
count of total and used bits is set to the number of *new* bits added by
the resize.  It should be the number of existing bits plus the number of
new bits.

The counter used to start at the number of existing bits, adding new
bits as offline resize went along.  Now it starts at zero, counting just
the new bits.

The solution is simple: when updating the global bitmap, add the new
counts to the existing ones.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>