tunefs.ocfs2: Create tunefs_main() and tunefs_feature_main()
tunefs_main() can be called from a method as the main() function.  It
takes a struct tunefs_operation, which provides the callbacks necessary
to perform that operation.

tunefs_feature main() can be called from a feature method as the main()
function.  It takes a struct tunefs_feature, which provides the callbacks
necessary to enable and disable a feature.

Signed-off-by: Joel Becker <joel.becker@oracle.com>