libocfs2: Make feature strings a proper API
The feature string functions are used in multiple places.  They belong
as a proper API in ocfs2.h  As such, move the required prototypes there,
give the functions and structures the appropriate prefixes, and remove
feature_string.h.

In addition, a DEBUG_EXE section is added to test the functions.  It
parses feature options and prints what mkfs and tunefs would do with them.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>