libocfs2: Clearing features should clear their dependancies.
The ocfs2_parse_feature() API was not symmetric.  If asked to set a
feature, it would set the feature and its dependencies.  However, when
asked to clear a feature, it would not clear the other features that
depended on it.  Thus, a caller looking at the list of features to clear
would have to know the reverse dependencies as well.

This change modifies ocfs2_parse_feature() to include the reverse
dependencies.  In addition, ocfs2_merge_feature_with_level() takes
advantage of the change to become simpler.

Finally, tunefs.ocfs2 now expects the reverse dependencies.  Instead of
adding them by hand, it makes sure they exist and behaves appropriately.

Signed-off-by: Joel Becker <joel.becker@oracle.com>