ocfs2_controld: genericize ocfs2_controld <-> cman interface
Fix up the Makefile to make it easier to plug in a new stack. Also, give
the exported functions in cman.c more generic names.

Aside from Makefile / configure.in changes, all that's required to support a
new stack in ocfs2_controld is an ocfs2_controld/stackname.c file which
provides:

        int setup_stack(void);
        char *nodeid2name(int nodeid);
        int validate_cluster(const char *cluster);
        int get_clustername(const char **cluster);
        int kill_stack_node(int nodeid);
        void exit_stack(void);

Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>