tunefs rework: Add op_set_slot_count.c
The set_slot_count operation changes the number of node slots on an
ocfs2 filesystem.  Each slot has a journal and other associated files.
The number of slots is the number of nodes that can share the
filesystem.  All filesystems must have at least one slot.

This is the tunefs option '-N'.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>