fsck.ocfs2: Reinitialize o2fsck_state when we reset fs_blocks.
In fsck.ocfs2, when we find fs_blocks in super block isn't the
same as the number stored in global_bitmap, we choose to trust
global_bitmap. In this case, we have to reinitialize the bitmap
in o2fsck_state since they are fs_blocks based. This bug blocks
the test case of fsck-test.sh in ocfs2-test.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>