ocfs2_controld: pacemaker support
This adds the build changes, and the core support for the pacemaker cluster
project.

Pacemaker is a next generation cluster resource manager, formerly part of
the heartbeat project. For more details, see:

http://www.clusterlabs.org/

Since pacemaker uses the openais api's, integration is fairly simple - we only
need drop in a pacemaker.c file which fills out the required ocfs2_controld
api's.

[ Whitespace, coding style fixes, naming fixes --Mark ]

Signed-off-by: Andrew Beekhof <abeekhof@suse.de>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>