ocfs2_controld,debugfs.ocfs2: Dump the daemon debug buffer via debugfs.ocfs2.
ocfs2_controld maintains a circular buffer of debug messages.  Until
now, the only way to see it was to run in the foreground via the -D
switch.  That's annoying to have to set up, and not useful in a
production environment.

We introduce the DUMP message to the controld protocol.  This message
requests the entire debug buffer.  ocfs2_controld sends the buffer as a
LIST reply.  Each ITEM in the list is one OCFS2_CONTROLD_MAXLINE packet.
Thus, the ITEMCOUNT is (roughly) "size of dump buffer /
OCFS2_CONTROLD_MAXLINE".

libo2cb gains the o2cb_control_daemon_debug() API.  This function sends
the DUMP message, receives the list, and returns a single string with
the entire buffer concatenated back together.

debugfs.ocfs2 gains the "controld" command.  Currently it only has one
sub-operation, "dump".  This calls o2cb_control_daemon_debug(), then
prints the results via the pager.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>