ocfs2_controld: Handle the corosync/openais split.
The openais project has split into two projects.  corosync is the core
totem infrastructure, and openais is the AIS API on top of it.  We need
to handle both pre- and post-split installations.

configure.in learns how to find libcpg in both places, setting
HAVE_COROSYNC when post-split.  ocfs2_controld/cpg.c includes the right
file based on HAVE_COROSYNC.

Signed-off-by: Joel Becker <joel.becker@oracle.com>