ocfs2_controld: Fail-fast when leaving a mountgroup unexpectedly.
When ocfs2_controld leaves a group unexpectedly (notified from cpg,
unclean exit), it currently sends node down events to the filesystem.
This doesn't work, because the filesystem BUG()s when given the local
node number.

A better solution is to fail-fast.  Exit right when we discover we are
leaving a live group unexpectedly.  If the group isn't live, clean up
our state and continue.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>