fsck.ocfs2: Move to JBD2 header and structures.
Change all defines and structures to JBD2 structures.  This is backwards
compatible with the JBD journals, and is capable of handling even the
oldest ocfs2 filesystems.

The only major functional change is to use journal_tag_size() instead of
sizeof(journal_block_tag_t) when walking the descriptor block.
fsck.ocfs2 and debugfs.ocfs2 are updated.

We deliberately put JBD2_FEATURE_INCOMPAT_REVOKE in the set of known
features.  We've never checked the feature set (we're about to), but
we've always supported revoke.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>