libocfs2: Add features to ocfs2_make_journal()
When we create/modify journals, we basically rewrite them.  The
ocfs2_make_journal() function now takes the set of features we want in
the journal.

mkfs.ocfs2 and fsck.ocfs2 are updated but only sets an empty feature
set.

tunefs.ocfs2 now checks the feature bits and is smart enough to copy the
feature set from other journals.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>