fsck.ocfs2: Fix feature bits, handle bad first journal.
fsck.ocfs2 needs to make the feature bits consistent across all
journals.

While we're there, we can fix the fact that fsck.ocfs2 implicitly trusts
the first journal file.  If the first journal is corrupt, it exits and
refuses to repair the filesystem.

Instead, we create a two-pass journal check scheme.  In the first pass,
all journals are checked.  If we can find a sane disposition from one of
them, we can then go about fixing all of them.

Included in this patch is fixing journal superblock feature bits.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>