tunefs.ocfs2: Add the cloned_volume operation.
Users with SANs often want to create clones or snapshots of ocfs2 volumes.
However, these volumes cannot be mounted, because the original volume
has the same UUID and is still in use.

The cloned_volume operation will reset the UUID and label of the given
device.  It cannot check with the cluster stack, so it must call
tunefs_interact_critical().

The --cloned-volume option has an optional <new-label> argument.  If
specified, the volume's label will be set to <new-label>.  If not, the
current volume label with have '-cloned' appended to it.  If the label
already ends in '-cloned', it is not modified.  The label will be
truncated to fit '-cloned'.

Because op_cloned_volume cannot start the cluster stack, a new
tunefs_open() flag is provided: TUNFES_FLAG_SKIPCLUSTER.  This skips all
cluster initialization, including dirty journal checks (which can't be
fixed without the cluster).  Operations using TUNEFS_FLAG_SKIPCLUSTER
are run before operations specifying TUNEFS_FLAG_NOCLUSTER.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>