fsck.ocfs2: Add replay orphan dir in slot recovery.
In o2fsck_pass4, we replay orphan dir. Now in slot recovery, we also
want to replay it, so export it from pass4. There is only one diffrence.
In slot recovery, we should exit the replay process if we meet with any
error, while in pass4, we should go on the process.
We also need to clear the link count of these dirs in the end during
slot recovery.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>