libocfs2: Make allocating a specific cluster a clean API
The backup superblock code hand-builds the necessary operations to load
the cluster allocator and set the backup superblock clusters.  Let's
create clean APIs for this in libocfs2/alloc.c and have backup_super.c
use them.

While we're at it, set_backup_super() used to free the cluster allocator
when it was done.  That doesn't fit with the library's standard "cache
that allocator" behavior.  So don't free it.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>