ocfs2-tools: ocfs2_controld leaves std(in|out|err) unassigned
Leaving std(in|out|err) unassigned causes havoc when included code (in
this case openais libraries) print to std(err|out).
Specifically, this often leads to memory corruption since the output
is sent to IPC sockets.

The safest approach is to reopen std(in|out|err) and have them point
to /dev/null

Signed-off-by: Andrew Beekhof <abeekhof@suse.de>
Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>