mkfs: Make 8 slots the default for most volume sizes
Patch updates the code that determines the default number of slots
for a volume. It now chooses 8 slots for volumes ranging from 10G to 1T.
Earlier the tool used to default to 4 slots. This change is long overdue
as the average size of the clusters is now greater than 4 nodes.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>