fsck.ocfs2(1.2): Modify e_clusters if it isn't equal to eb contains.
In bug 7553939, we meet with a scenario that the clusters haven't
been allocated to the extent block(no update) while the inode already
have the i_clusters and e_clusters updated. The commit
09166ad03c83f39f2dcd77c74cd4672c3a371dfa has fixed the problem
of i_clusters corruption, so this one intends to fix e_clusters.
We calculate the total clusters within the extent block and then
modify the extent record if it doesn't show the right value.

Please note that to be safe, it only fix the corruption happens in the
last extent block. If there is any corruption happens in other extent block,
we don't touch it(actually if we modify it, we may leave a hole there and
make the system corrupted). The good thing is that in ocfs2-1.2, we don't
support sparse files, and we only extend files, so if there is one corruption
like 7553939, it must happen in the last extent block.

One more thing, in ocfs2-1.4, we support sparse files, so it can be handled
gracefully. So this is only a fix for ocfs2-tools-1.2.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>