mkfs.ocfs2: Update fs-features list in man page
The list of available options for the fs-features flag in mkfs.ocfs2.8 was
out-dated. Update the list by added information on the 'backup-super',
'inline-data' and 'extended-slotmap' options. Also, update the reported list
of set features for the --fs-feature-level=default option.

At the same time, we remove the even more out of date list from the
tunefs.ocfs2 man page and provide a pointer to the mkfs.ocfs2 man page. This
way we only have one place to update.

Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>