libocfs2: Add the in-progress flag for enabling dirblock trailers.
When tunefs.ocfs2 is adding directory block trailers, it cannot do it
atomically.  For some directory blocks, dirents have to be moved out of
the way to make room for the trailer.  These dirents are shifted to new
directory blocks.

When tunefs writes out the changes, it does so in the following order:

1) Write out the new blocks
2) Update the directory's i_size to include the new blocks
3) Write the old blocks with trailers in place.

If tunefs crashes after (1), the new blocks are just unused allocation.
We're safe.  If tunefs crashes after (3), we have the directory with
trailers correctly installed.  However, if tunefs crashes after (2), the
old dirents are in the old blocks, but the shifted copies are in the new
blocks.  This leaves duplicate dirents.  fsck.ocfs2 can fix them, but we
need to set OCFS2_TUNEFS_INPROG_DIR_TRAILER to make sure fsck is run.

The alternative order, writing out the old blocks before the new ones,
leaves missing dirents in the case of a crash.  That's not a good plan.

Signed-off-by: Joel Becker <joel.becker@oracle.com>