tunefs.ocfs2: Enable and disable the metaecc feature.
The metaecc feature requires a bit of work.

To disable it is easy; just clear the feature bit.  All structures left
behind are fully compatible with filesystems that don't know about
metaecc.

Enabling it is hard work.  We must run though every block in the
filesystem and compute its checksum and error correction values.  But
even that's not all.  Directory blocks need trailers to store the ECC
values, and the directories might not have them yet.  More on the
trailers below.

The preparation step runs through every inode in the filesystem.  If the
inode does not need directory trailers - that is, it is not a directory,
is an inline directory, or already has trailers - we add all of its
metadata blocks to an rbtree.  This caches the blocks for later
re-write.  We don't want to have to re-read them in the writeout step.
If the inode is a directory requiring trailers, we call
tunefs_prepare_dir_trailer() to determine how much space it needs.  At
the end of this step, we have a list of all directories requiring
trailers, and we know whether we have enough space to do the work.

Next, we install all the trailers.  When this completes, we have a valid
filesystem with trailers.  For each directory, we add its blocks to the
rbtree after we've added the trailers.

Finally, we run through the rbtree and compute the ECC values for each
block.  When this is finished, we can write the META_ECC feature bit to
the superblock.

We add two functions to handle the trailers.

tunefs_prepare_dir_trailer() will walk a directory and check all of its
dirblocks.  If the directory needs trailers, a tunefs_trailer_context is
returned.  This contains all the information needed to add trailers to
this directory.  The caller can use this structure to determine how much
allocation is needed.  The caller passes this structure to
tunefs_install_dir_trailer() to actually write out the trailers for this
directory.

We add a tunefs-in-progress flag for adding dirblock trailers.

When tunefs.ocfs2 is adding directory block trailers, it cannot do it
atomically.  For some directory blocks, dirents have to be moved out of
the way to make room for the trailer.  These dirents are shifted to new
directory blocks.  When tunefs writes out the changes, it does so in the
following order:

1) Write out the new blocks
2) Update the directory's i_size to include the new blocks
3) Write the old blocks with trailers in place.

If tunefs crashes after (1), the new blocks are just unused allocation.
We're safe.  If tunefs crashes after (3), we have the directory with
trailers correctly installed.  However, if tunefs crashes after (2), the
old dirents are in the old blocks, but the shifted copies are in the new
blocks.  This leaves duplicate dirents.  fsck.ocfs2 can fix them, but we
need to set OCFS2_TUNEFS_INPROG_DIR_TRAILER to make sure fsck is run.

The alternative order, writing out the old blocks before the new ones,
leaves missing dirents in the case of a crash.  That's not a good plan.

Signed-off-by: Joel Becker <joel.becker@oracle.com>